+++ include/library/pixfmt-rgb32.hpp
@@ -10,7 +10,7 @@ class pixel_format_rgb32 : public pixel_format
 {
 public:
 	~pixel_format_rgb32() throw();
-	void decode(uint8_t* target, const uint8_t* src, size_t width)
+	void decode(uint32_t* target, const uint8_t* src, size_t width)
 		throw();
 	void decode(uint32_t* target, const uint8_t* src, size_t width,
 		const pixel_format_aux_palette<false>& auxp) throw();
